+++ CHANGELOG.md
@@ -2,6 +2,14 @@
 
 All notable changes to `laravel-aws-cloud-watch-logger` will be documented in this file
 
-## 0.0.1 - 2021-08-16
+## 0.0.2 - 2021-08-19
+
+- file and cloudwatch drive is now ready to use
+- Publish and replace capability added
+- Return type changed
+- phpDocs added for Facade auto discovery
+- function visibility changed
+
+## 0.0.1 - 2021-08-18
 
 - initial release